+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/LoadTest.kt
@@ -5,7 +5,7 @@ import com.atlassian.performance.tools.virtualusers.api.VirtualUserNodeResult
 import com.atlassian.performance.tools.virtualusers.api.VirtualUserOptions
 import com.atlassian.performance.tools.virtualusers.api.config.LoadProcessContainer
 import com.atlassian.performance.tools.virtualusers.api.config.LoadThreadContainer
-import com.atlassian.performance.tools.virtualusers.engine.LoadThread
+import com.atlassian.performance.tools.virtualusers.api.load.LoadThread
 import java.util.*
 import java.util.concurrent.LinkedBlockingQueue
 import java.util.concurrent.ThreadPoolExecutor
@@ -16,10 +16,10 @@ import java.util.concurrent.atomic.AtomicInteger
 internal class LoadTest(
     private val options: VirtualUserOptions
 ) {
-    private val behavior = options.behavior
-    private val process = behavior.loadProcess.getConstructor().newInstance()
 
     fun run(): VirtualUserNodeResult {
+        val behavior = options.behavior
+        val process = behavior.loadProcess.getConstructor().newInstance()
         val processContainer = LoadProcessContainer.create(
             options,
             VirtualUserNodeResult(behavior.results),
@@ -40,10 +40,10 @@ internal class LoadTest(
             Thread(runnable, name).apply { isDaemon = true }
         }
         val stop = AtomicBoolean(false)
-        val threadFactory = process.setUp(processContainer)
+        val threadFactory = process.prepareFactory(processContainer)
         val threads = (1..threadCount).map { threadIndex ->
             val threadContainer = LoadThreadContainer.create(processContainer, threadIndex, UUID.randomUUID())
-            val readyThread = threadFactory.fireUp(threadContainer)
+            val readyThread = threadFactory.prepareThread(threadContainer)
             ContainedThread(readyThread, threadContainer)
         }
         threads.forEach { engine ->