+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/load/LoadThreadFactory.kt
@@ -1,0 +1,12 @@
+package com.atlassian.performance.tools.virtualusers.api.load
+
+import com.atlassian.performance.tools.virtualusers.api.config.LoadThreadContainer
+import net.jcip.annotations.ThreadSafe
+
+/**
+ * Used multiple times per [LoadProcess].
+ */
+@ThreadSafe
+interface LoadThreadFactory {
+    fun prepareThread(container: LoadThreadContainer): LoadThread
+}