+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/engine/LoadThreadFactory.kt
@@ -1,7 +1,0 @@
-package com.atlassian.performance.tools.virtualusers.engine
-
-import com.atlassian.performance.tools.virtualusers.api.config.LoadThreadContainer
-
-interface LoadThreadFactory {
-    fun fireUp(container: LoadThreadContainer): LoadThread
-}