+++ src/test/kotlin/com/atlassian/performance/tools/virtualusers/api/VirtualUserOptionsTest.kt
@@ -9,9 +9,8 @@ import com.atlassian.performance.tools.virtualusers.api.config.VirtualUserTarget
 import com.atlassian.performance.tools.virtualusers.api.users.RestUserGenerator
 import com.atlassian.performance.tools.virtualusers.api.users.SuppliedUserGenerator
 import com.atlassian.performance.tools.virtualusers.api.users.UserGenerator
-import com.atlassian.performance.tools.virtualusers.engine.HttpLoadProcess
-import com.atlassian.performance.tools.virtualusers.engine.LoadProcess
-import com.atlassian.performance.tools.virtualusers.engine.LoadThreadFactory
+import com.atlassian.performance.tools.virtualusers.api.load.LoadProcess
+import com.atlassian.performance.tools.virtualusers.api.load.LoadThreadFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.catchThrowable
 import org.junit.Test
@@ -134,7 +133,7 @@ class VirtualUserOptionsTest {
             )
             .containsSequence(
                 "--scenario",
-                "com.atlassian.performance.tools.virtualusers.engine.DoNotUseWhenEngineIsProvided"
+                "com.atlassian.performance.tools.virtualusers.load.DoNotUseWhenLoadProcessIsProvided"
             )
     }
 
@@ -147,11 +146,11 @@ class VirtualUserOptionsTest {
         assertThat(options.toCliArgs())
             .containsSequence(
                 "--load-process",
-                "com.atlassian.performance.tools.virtualusers.engine.HttpLoadProcess"
+                "com.atlassian.performance.tools.virtualusers.load.HttpLoadProcess"
             )
             .containsSequence(
                 "--scenario",
-                "com.atlassian.performance.tools.virtualusers.engine.DoNotUseWhenEngineIsProvided"
+                "com.atlassian.performance.tools.virtualusers.load.DoNotUseWhenLoadProcessIsProvided"
             )
     }
 
@@ -299,7 +298,7 @@ class VirtualUserOptionsTest {
     }
 
     private class MockLoadProcess : LoadProcess {
-        override fun setUp(container: LoadProcessContainer): LoadThreadFactory {
+        override fun prepareFactory(container: LoadProcessContainer): LoadThreadFactory {
             throw Exception("no need to call it in this test")
         }
     }