+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.14.1...master
+[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.15.0...master
+
+## [3.15.0] - 2023-09-21
+[3.15.0]: https://github.com/atlassian/virtual-users/compare/release-3.14.1...release-3.15.0
 
 ### Added
 - Add `LoadProcess`, `LoadThreadFactory` and `LoadThread` SPIs for greater control over load generation.